Move field DSL methods to module, to keep core Field stuff easy to follow.
